@@ -4,7 +4,7 @@ description: Showcase for isolate_bloc architecture pattern.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -13,8 +13,8 @@ dependencies:
   cupertino_icons: ^1.0.0
   isolate_bloc:
     path: ../../packages/isolate_bloc/
-  equatable: any
-  http: any
+  equatable: 2.0.3
+  http: 0.13.3
 
 dev_dependencies:
   flutter_test:
