@@ -1,15 +1,17 @@
 name: isolate_bloc_test
 description: A testing library which make it easy to test isolate_bloc.
-version: 0.1.1
+version: 2.0.0
 homepage: https://github.com/Maksimka101/isolate-bloc/tree/master/packages/isolate_bloc_test
 repository: https://github.com/Maksimka101/isolate-bloc/tree/master/packages/isolate_bloc_test
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0 <3.0.0"
 
 dependencies:
-  isolate_bloc: ^0.4.0
+  # isolate_bloc: ^2.0.0
+  isolate_bloc: 
+    path: ../isolate_bloc/
   flutter:
     sdk: flutter
 
