@@ -1,7 +1,6 @@
 import 'dart:convert';
 import 'dart:async';
 
-import 'package:meta/meta.dart';
 import 'package:http/http.dart' as http;
 
 import '../models/weather.dart';
@@ -10,11 +9,11 @@ class WeatherApiClient {
   static const baseUrl = 'https://www.metaweather.com';
   final http.Client httpClient;
 
-  WeatherApiClient({@required this.httpClient}) : assert(httpClient != null);
+  WeatherApiClient({required this.httpClient});
 
   Future<int> getLocationId(String city) async {
     final locationUrl = '$baseUrl/api/location/search/?query=$city';
-    final locationResponse = await this.httpClient.get(locationUrl);
+    final locationResponse = await this.httpClient.get(Uri.parse(locationUrl));
     if (locationResponse.statusCode != 200) {
       throw Exception('error getting locationId for city');
     }
@@ -25,7 +24,7 @@ class WeatherApiClient {
 
   Future<Weather> fetchWeather(int locationId) async {
     final weatherUrl = '$baseUrl/api/location/$locationId';
-    final weatherResponse = await this.httpClient.get(weatherUrl);
+    final weatherResponse = await this.httpClient.get(Uri.parse(weatherUrl));
 
     if (weatherResponse.statusCode != 200) {
       throw Exception('error getting weather for location');
