@@ -1,13 +1,10 @@
 import 'package:flutter/material.dart';
 
-import 'package:meta/meta.dart';
 
 class Location extends StatelessWidget {
   final String location;
 
-  Location({Key key, @required this.location})
-      : assert(location != null),
-        super(key: key);
+  Location({Key? key, required this.location}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
