@@ -1,14 +1,11 @@
 import 'package:flutter/material.dart';
 
-import 'package:meta/meta.dart';
 import 'package:weather_app/models/models.dart';
 
 class WeatherConditions extends StatelessWidget {
   final WeatherCondition condition;
 
-  WeatherConditions({Key key, @required this.condition})
-      : assert(condition != null),
-        super(key: key);
+  WeatherConditions({Key? key, required this.condition}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => _mapConditionToImage(condition);
