@@ -1,11 +1,12 @@
-import '../bloc/isolate_bloc.dart';
-import '../bloc/isolate_bloc_wrapper.dart';
-import 'isolate_connector.dart';
-import 'isolate_manager/abstract_isolate_manager.dart';
-import 'isolate_manager/isolate_messenger.dart';
-import 'isolated_bloc_manager.dart';
-import 'isolated_connector.dart';
-import 'service_events.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_base.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_wrapper.dart';
+import 'package:isolate_bloc/src/common/isolate/isolate_connector.dart';
+import 'package:isolate_bloc/src/common/isolate/isolate_manager/abstract_isolate_manager.dart';
+import 'package:isolate_bloc/src/common/isolate/isolate_manager/isolate_messenger.dart';
+import 'package:isolate_bloc/src/common/isolate/isolated_bloc_manager.dart';
+import 'package:isolate_bloc/src/common/isolate/isolated_connector.dart';
+import 'package:isolate_bloc/src/common/isolate/platform_channel/platform_channel_setup.dart';
+import 'package:isolate_bloc/src/common/isolate/service_events.dart';
 
 /// Signature for initialization function which would be run in [Isolate] to
 /// initialize your blocs and repository.
@@ -14,38 +15,39 @@ typedef Initializer = Function();
 
 typedef IsolateManagerCreator = Future<IsolateManager> Function(
   IsolateRun,
-  Initializer, [
+  Initializer,
   List<String> channels,
-]);
+);
 
 /// Maintain [IsolateBlocWrapper]s
 class BlocManager {
   BlocManager._(
-      this._initialStates, this._isolateConnector, this._isolateManager);
+    this._initialStates,
+    this._isolateConnector,
+    this._isolateManager,
+  );
 
-  static BlocManager instance;
+  static BlocManager? instance;
 
   final IsolateConnector _isolateConnector;
   final IsolateManager _isolateManager;
-  final Map<Type, Object> _initialStates;
+  final Map<Type, Object?> _initialStates;
   final _freeWrappers = <Type, List<IsolateBlocWrapper>>{};
   final _wrappers = <String, IsolateBlocWrapper>{};
 
-  /// Create [Isolate], run user [Initializer] and perform other tasks that are
+  /// Creates [Isolate], run user's [Initializer] and perform other tasks that are
   /// necessary for this library to work.
   /// Ensure that your app or interactions with this library starts with this
   /// function.
-  /// If already initialized kill previous [Isolate] and reinitialize everything.
+  /// If already initialized kills previous [Isolate] and reinitialize.
   static Future<void> initialize(
     Initializer initializer,
-    IsolateManagerCreator createIsolate,
-    List<String> platformChannels,
+    IsolateManagerFactory managerFactory,
+    MethodChannels platformChannels,
   ) async {
-    if (instance != null) {
-      instance.dispose();
-    }
+    instance?.dispose();
 
-    final isolateManager = await createIsolate(
+    final isolateManager = await managerFactory.create(
       _isolatedBlocRunner,
       initializer,
       platformChannels,
@@ -66,36 +68,45 @@ class BlocManager {
     );
   }
 
-  /// Start creating [IsolateBloc] and return [IsolateBlocWrapper].
-  IsolateBlocWrapper<State> createBloc<T extends IsolateBloc, State>() {
+  /// Start creating [IsolateBlocBase] and return [IsolateBlocWrapper].
+  IsolateBlocWrapper<State> createBloc<T extends IsolateBlocBase, State>() {
+    void onBlocClose(String? uuid) {
+      if (uuid != null) {
+        _isolateConnector.sendEvent(CloseIsolateBlocEvent(uuid));
+      }
+    }
+
     final initialState = _initialStates[T];
     final messageReceiver = _isolateConnector.sendEvent;
-    final onBlocClose = (String uuid) =>
-        _isolateConnector.sendEvent(CloseIsolateBlocEvent(uuid));
+
     final blocWrapper = IsolateBlocWrapper<State>(
-        initialState as State, messageReceiver, onBlocClose);
-    if (!_freeWrappers.containsKey(T)) {
-      _freeWrappers[T] = [];
-    }
-    _freeWrappers[T].add(blocWrapper);
+      state: initialState as State,
+      eventReceiver: messageReceiver,
+      onBlocClose: onBlocClose,
+    );
+
+    _freeWrappers[T] ??= [];
+    _freeWrappers[T]!.add(blocWrapper);
     _isolateConnector.sendEvent(CreateIsolateBlocEvent(T));
     return blocWrapper;
   }
 
-  /// Finish [IsolateBloc] creating which started by call [createBloc].
-  /// Connect [IsolateBloc] to it's [IsolateBlocWrapper].
+  /// Finish [IsolateBlocBase] creating which started by call [createBloc].
+  /// Connect [IsolateBlocBase] to it's [IsolateBlocWrapper].
   void bindFreeWrapper(Type blocType, String id) {
-    assert(
-      _freeWrappers.containsKey(blocType) && _freeWrappers[blocType].isNotEmpty,
-      'No free bloc wrapper for $blocType',
-    );
-    _wrappers[id] = _freeWrappers[blocType].removeAt(0)..connectToBloc(id);
+    if (_freeWrappers.containsKey(blocType) && _freeWrappers[blocType]!.isNotEmpty) {
+      // ignore: invalid_use_of_protected_member
+      _wrappers[id] = _freeWrappers[blocType]!.removeAt(0)..connectToBloc(id);
+    } else {
+      throw Exception('No free bloc wrapper for $blocType');
+    }
   }
 
-  /// Call when new state from [IsolateBloc] received.
+  /// Call when new state from [IsolateBlocBase] received.
   /// Find wrapper by bloc id and add new state to it.
-  void blocStateReceiver(String blocId, Object state) {
-    _wrappers[blocId].stateReceiver(state);
+  void blocStateReceiver(String blocId, Object? state) {
+    // ignore: invalid_use_of_protected_member
+    _wrappers[blocId]?.stateReceiver(state);
   }
 
   static Future<void> _isolatedBlocRunner(
@@ -106,7 +117,8 @@ class BlocManager {
       IsolatedConnector(
         messenger.add,
         Stream.castFrom<Object, ServiceEvent>(
-            messenger.where((event) => event is ServiceEvent)),
+          messenger.where((event) => event is ServiceEvent),
+        ),
       ),
     );
 
@@ -121,7 +133,7 @@ class BlocManager {
     isolateBlocManager.initializeCompleted();
   }
 
-  /// Free all resources and kill [Isolate] with [IsolateBloc]s.
+  /// Free all resources and kill [Isolate] with [IsolateBlocBase]s.
   void dispose() {
     _isolateManager.isolate.kill();
     _isolateConnector.dispose();
