@@ -2,7 +2,7 @@ import 'package:isolate_bloc/isolate_bloc.dart';
 
 class SimpleBlocObserver extends IsolateBlocObserver {
   @override
-  void onEvent(IsolateBloc bloc, Object event) {
+  void onEvent(IsolateBlocBase bloc, Object? event) {
     super.onEvent(bloc, event);
     print('onEvent $event');
   }
@@ -14,7 +14,7 @@ class SimpleBlocObserver extends IsolateBlocObserver {
   }
 
   @override
-  void onError(IsolateBloc bloc, Object error, StackTrace stackTrace) {
+  void onError(IsolateBlocBase bloc, Object error, StackTrace stackTrace) {
     super.onError(bloc, error, stackTrace);
     print('onError $error');
   }
