@@ -1,7 +1,7 @@
 import 'package:http/http.dart';
 import 'package:isolate_bloc/isolate_bloc.dart';
-import 'package:weather_app/blocs/settings_bloc.dart';
-import 'package:weather_app/blocs/theme_bloc.dart';
+import 'package:weather_app/blocs/settings_cubit.dart';
+import 'package:weather_app/blocs/theme_cubit.dart';
 import 'package:weather_app/blocs/weather_bloc.dart';
 import 'package:weather_app/repositories/repositories.dart';
 
@@ -16,9 +16,9 @@ void isolateInitialization() {
   );
 
   register(create: () => WeatherBloc(weatherRepository: weatherRepository));
-  register(create: () => SettingsBloc());
+  register(create: () => SettingsCubit());
   register(
-    create: () => ThemeBloc(
+    create: () => ThemeCubit(
       weatherBloc: getBloc<WeatherBloc, WeatherState>(),
     ),
   );
