@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:isolate_bloc/isolate_bloc.dart';
-import 'package:weather_app/blocs/theme_bloc.dart';
+import 'package:weather_app/blocs/theme_cubit.dart';
 import 'package:weather_app/blocs/weather_bloc.dart';
 import 'package:weather_app/utils/theme_utils.dart';
 import 'package:weather_app/widgets/widgets.dart';
@@ -13,7 +13,7 @@ class Weather extends StatefulWidget {
 }
 
 class _WeatherState extends State<Weather> {
-  Completer<void> _refreshCompleter;
+  Completer<void>? _refreshCompleter;
 
   @override
   void initState() {
@@ -56,7 +56,7 @@ class _WeatherState extends State<Weather> {
         ],
       ),
       body: Center(
-        child: IsolateBlocListener<ThemeBloc, ThemeState>(
+        child: IsolateBlocListener<ThemeCubit, ThemeState>(
           listener: (context, state) {
             _refreshCompleter?.complete();
             _refreshCompleter = Completer();
@@ -69,7 +69,7 @@ class _WeatherState extends State<Weather> {
               if (state is WeatherLoadSuccess) {
                 final weather = state.weather;
 
-                return IsolateBlocBuilder<ThemeBloc, ThemeState>(
+                return IsolateBlocBuilder<ThemeCubit, ThemeState>(
                   builder: (context, themeState) {
                     return GradientContainer(
                       color: mapThemeStateToColor(themeState),
@@ -80,7 +80,7 @@ class _WeatherState extends State<Weather> {
                               .add(
                             WeatherRefreshRequested(city: weather.location),
                           );
-                          return _refreshCompleter.future;
+                          return _refreshCompleter!.future;
                         },
                         child: ListView(
                           children: <Widget>[
