@@ -1,23 +1,21 @@
 import 'dart:async';
+import 'dart:isolate';
 
+import 'package:isolate_bloc/src/common/isolate/bloc_manager.dart';
 import 'package:isolate_bloc/src/common/isolate/isolate_manager/abstract_isolate_manager.dart';
-import 'package:isolate_bloc/src/common/isolate/isolate_manager/abstract_isolate_wrapper.dart';
 import 'package:isolate_bloc/src/common/isolate/isolate_manager/isolate_messenger.dart';
+import 'package:isolate_bloc/src/common/isolate/isolate_manager/web/web_isolate_wrapper.dart';
+import 'package:isolate_bloc/src/common/isolate/platform_channel/platform_channel_setup.dart';
 
-import '../../bloc_manager.dart';
-import 'web_isolate_wrapper.dart';
-
-/// Web [IsolateManager]'s implementation.
-/// It doesn't creates [Isolate].
-class IsolateManagerImpl extends IsolateManager {
-  IsolateManagerImpl(IsolateWrapper isolate, IsolateMessenger messenger)
-      : super(isolate, messenger);
-
-  static Future<IsolateManagerImpl> createIsolate(
+/// Web [IsolateManagerFactory]'s implementation.
+/// It doesn't create [Isolate].
+class WebIsolateManagerFactory implements IsolateManagerFactory {
+  @override
+  Future<IsolateManager> create(
     IsolateRun run,
-    Initializer initializer, [
-    List<String> platformChannels,
-  ]) async {
+    Initializer initializer,
+    MethodChannels methodChannels,
+  ) async {
     final fromIsolate = StreamController<Object>.broadcast();
     final toIsolate = StreamController<Object>.broadcast();
     final sendFromIsolate = fromIsolate.add;
@@ -30,7 +28,7 @@ class IsolateManagerImpl extends IsolateManager {
     // this function run isolated function (IsolateRun)
     run(IsolateMessenger(toIsolateStream, sendFromIsolate), initializer);
 
-    return IsolateManagerImpl(
+    return IsolateManager(
       WebIsolateWrapper(),
       isolateMessenger,
     );
