@@ -1,20 +1,22 @@
-import '../common/bloc/isolate_bloc.dart';
-import '../common/bloc/isolate_bloc_wrapper.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_base.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_wrapper.dart';
 
-/// Use it to store bloc uuid's and type's in context using provider.
+/// Use it to store bloc in context using provider.
 class BlocInfoHolder {
   final _blocsInfo = <Type, IsolateBlocWrapper>{};
 
-  /// Return [IsolateBlocWrapper] associated with given [IsolateBloc]'s Type
-  IsolateBlocWrapper<State>
-      getWrapperByType<T extends IsolateBloc<Object, State>, State>() =>
-          _blocsInfo[T] as IsolateBlocWrapper<State>;
+  /// Return [IsolateBlocWrapper] associated with given [IsolateCubit]'s Type
+  IsolateBlocWrapper<State>? getWrapperByType<T extends IsolateBlocBase<Object?, State>, State>() {
+    return _blocsInfo[T] as IsolateBlocWrapper<State>?;
+  }
 
-  /// Add [IsolateBlocWrapper] associated with [IsolateBloc] type.
-  void addBlocInfo<T extends IsolateBloc>(IsolateBlocWrapper wrapper) =>
-      _blocsInfo[T] = wrapper;
+  /// Add [IsolateBlocWrapper] associated with [IsolateCubit] type.
+  void addBlocInfo<T extends IsolateBlocBase>(IsolateBlocWrapper wrapper) {
+    _blocsInfo[T] = wrapper;
+  }
 
-  /// Remove [IsolateBlocWrapper] associated with [IsolateBloc]
-  IsolateBlocWrapper removeBloc<T extends IsolateBloc>() =>
-      _blocsInfo.remove(T);
+  /// Remove [IsolateBlocWrapper] associated with [IsolateCubit]
+  IsolateBlocWrapper? removeBloc<T extends IsolateBlocBase>() {
+    _blocsInfo.remove(T);
+  }
 }
