@@ -1,7 +1,7 @@
 import 'package:isolate_bloc/isolate_bloc.dart';
 
 /// Simple bloc which is send "empty" as initial state and "data" as any another states.
-class SimpleBloc extends IsolateBloc<Object, String> {
+class SimpleBloc extends IsolateCubit<Object, String> {
   SimpleBloc() : super("empty");
 
   @override
