@@ -5,28 +5,22 @@ import 'package:isolate_bloc/isolate_bloc.dart';
 import 'mock_isolate_wrapper.dart';
 
 /// [IsolateManager] implementation for tests.
-/// Have all [IsolateManagerImpl]'s restrictions.
-class MockIsolateManager extends IsolateManager {
-  MockIsolateManager(IsolateWrapper isolate, IsolateMessenger messenger)
-      : super(isolate, messenger);
-
-  /// Create [MockIsolateManager] object.
-  static Future<MockIsolateManager> createIsolate(
-    IsolateRun run,
-    Initializer initializer, [
-    List<String> platformChannels,
-  ]) async {
+/// Have all IO [IsolateManager] restrictions.
+class MockIsolateManagerFactory extends IsolateManagerFactory {
+  /// Create mock [IsolateManager] object.
+  @override
+  Future<IsolateManager> create(IsolateRun isolateRun, Initializer initializer, MethodChannels methodChannels) async {
     var fromIsolate = ReceivePort();
     var toIsolate = ReceivePort();
     var sendFromIsolate = fromIsolate.sendPort.send;
     var sendToIsolate = toIsolate.sendPort.send;
-    var toIsolateStream = toIsolate.asBroadcastStream();
-    var fromIsolateStream = fromIsolate.asBroadcastStream();
+    var toIsolateStream = toIsolate.asBroadcastStream().cast<Object>();
+    var fromIsolateStream = fromIsolate.asBroadcastStream().cast<Object>();
 
     // this function run isolated function (IsolateRun)
-    run(IsolateMessenger(toIsolateStream, sendFromIsolate), initializer);
+    isolateRun(IsolateMessenger(toIsolateStream, sendFromIsolate), initializer);
 
-    return MockIsolateManager(
+    return IsolateManager(
       MockIsolateWrapper(),
       IsolateMessenger(
         fromIsolateStream,
